pyproject.toml
[tool.poetry]
name = "repl_python3_TotalQuietOutsourcing"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.19.2"
urllib3 = "^1.26.7"
requests = "^2.26.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
